pyproject.toml
[project]
name = "rewear"
version = "1.0.0"
description = "ReWear - Community Clothing Exchange Platform"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "pillow>=11.3.0",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
